Multiple Files: move load_system_configuration from utils to system_settings; update all references; update unittest
